[CBRD-27140] Widen row and key counts in the statistics chain to INT64
Row counts (CLS_INFO.ci_tot_objects, CLASS_STATS.heap_num_objects), index
key counts and partial-key counts (BTREE_STATS.keys/pkeys, the optimizer's
QO_ATTR_CUM_STATS) were 32-bit end to end: the b-tree key counters wrapped
negative past 2^31 keys, the wrapped value was persisted in the catalog and
shipped to the client, the optimizer then dropped the index NDV altogether,
and debug builds aborted in btree_get_stats().

- statistics.h / system_catalog.h / optimizer.h: INT64 fields; the pkeys
  arrays are allocated as INT64 everywhere.
- btree.c: the Duj1 estimator and the prefix-skip-scan budget compute in
  INT64.
- statistics_sr.c / statistics_cl.c: INT64 on the wire for tot_objects,
  keys and pkeys[]; partition sums saturate through stats_double_to_int64()
  instead of a double->int cast (undefined past INT_MAX). The
  heap_get_num_objects() fallback now checks a real status code.
- system_catalog.c: DISK_REPR gains stats_layout in the formerly reserved
  slot. V0 (80-byte BTREE_STATS, 32-bit) records are still read; writers
  emit V1 (120 bytes, INT64 keys/pkeys). CLS_INFO stores the INT64 count in
  its zero-filled tail and falls back to the 32-bit field when that is 0.
- heap_file.c: heap_get_num_objects()/heap_estimate() return the count
  through an INT64 out-parameter; heap_get_num_objects() returns NO_ERROR.
- Consumers that stay int (db_get_btree_statistics(),
  db_get_class_num_objs_and_pages(), index_cardinality(), SHOW INDEX)
  saturate via STATS_CLAMP_TO_INT. The #objects/#keys pseudo columns become
  BIGINT.


# 제목: 변경 요약 (50자 이내)  (refactor, hotfix, feature)

# 본문: 변경 사항에 대한 자세한 설명
# - 주요 변경 사항 1
# - 주요 변경 사항 2

# 참고: 이 라인 아래의 내용은 커밋 메시지에서 제거됩니다.

+++ src/storage/heap_file.h
@@ -456,9 +456,10 @@ extern SCAN_CODE heap_scanrange_last (THREAD_ENTRY * thread_p, OID * last_oid, R
 extern bool heap_does_exist (THREAD_ENTRY * thread_p, OID * class_oid, const OID * oid);
 extern bool heap_is_object_not_null (THREAD_ENTRY * thread_p, OID * class_oid, const OID * oid);
 extern int heap_get_num_data_pages (THREAD_ENTRY * thread_p, const HFID * hfid, int *num_pages);
-extern int heap_get_num_objects (THREAD_ENTRY * thread_p, const HFID * hfid, int *npages, int *nobjs, int *avg_length);
+extern int heap_get_num_objects (THREAD_ENTRY * thread_p, const HFID * hfid, int *npages, INT64 * nobjs,
+				 int *avg_length);
 
-extern int heap_estimate (THREAD_ENTRY * thread_p, const HFID * hfid, int *npages, int *nobjs, int *avg_length);
+extern int heap_estimate (THREAD_ENTRY * thread_p, const HFID * hfid, int *npages, INT64 * nobjs, int *avg_length);
 extern int heap_estimate_num_objects (THREAD_ENTRY * thread_p, const HFID * hfid);
 
 extern int heap_get_class_name (THREAD_ENTRY * thread_p, const OID * class_oid, char **class_name);
